add standalone Talon runtime
