+++ bench/benchmark_1000_agents.py
@@ -199,7 +199,7 @@ def talon_command(database: str) -> str:
     if database == "rocksdb":
         return """
 set -eu
-exec talon-bench-colocated
+exec talon-standalone
 """.strip()
 
     return """
